exception handling

+++ src/main/java/com/giovaniwahl/dscommerce/domain/services/ProductService.java
@@ -3,11 +3,15 @@ package com.giovaniwahl.dscommerce.domain.services;
 import com.giovaniwahl.dscommerce.domain.dtos.ProductDTO;
 import com.giovaniwahl.dscommerce.domain.entities.Product;
 import com.giovaniwahl.dscommerce.domain.repositories.ProductRepository;
+import com.giovaniwahl.dscommerce.domain.services.exceptions.DatabaseException;
 import com.giovaniwahl.dscommerce.domain.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -41,14 +45,27 @@ public class ProductService {
     }
     @Transactional
     public ProductDTO update(Long id,ProductDTO productDTO){
-        Product product = productRepository.getReferenceById(id);
-        copyDtoToEntity(productDTO,product);
-        product = productRepository.save(product);
-        return new ProductDTO(product);
+        try {Product product = productRepository.getReferenceById(id);
+            copyDtoToEntity(productDTO,product);
+            product = productRepository.save(product);
+            return new ProductDTO(product);
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Resource not found!");
+        }
+
     }
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void  delete(Long id){
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Resource not found !");
+        }
+        try {
+            productRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Referential integrity failure !");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO productDTO, Product product) {
